Update help text & move account to separate file
